+++ src/DockerWatch/ContainerMonitorHostOptions.cs
@@ -1,0 +1,7 @@
+namespace DockerWatch
+{
+    public class ContainerMonitorHostOptions
+    {
+        public string ContainerGlob { get; set; }
+    }
+}