+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] - 2018-08-31
+### Added
+- New parameter (`--container`) that accepts a glob to filter the running containers for which you would like to watch for cahnges
+
+### Fixed
+- Monitor throws an error for volumes that only have a mount point in the container, but are not mounted in windows.
+
 ## [v0.2.0] - 2018-07-26
 ### Added
 - improved logging